[Android] add toplevel Android.mk to stop breakage
Summary:
Android is currently moving over to building AOSP's LLVM release from
the monorepo. In testing, we were hitting breakage due to the Android.mk
file under pstl/ (parallel STL). Previously, Android was not including
pstl/ as a separate repo and now with the monorepo we get it no matter
what (and will probably use it soon).

The issue for Android's build system is that it recursively looks for
Android.mk files, so it picks up the one under pstl/ even though we
don't want to build it just yet.

The sugguested fix from Android's build team is to add an empty
Android.mk to the top level directory, which will prevent the build
system from recursing further, finding the one in pstl/, and erroring
out.

Link: https://android-review.googlesource.com/c/platform/manifest/+/977993

Reviewers: srhines, danalbert, MikeDvorskiy

Subscribers: llvm-commits, pirama

Tags: #llvm

Differential Revision: https://reviews.llvm.org/D63017

Change-Id: I6d42ff21a6516bc73acb9a16094bd9e9e495e562