Update patch set 2
Patch Set 2:

> Patch Set 1:
> 
> Can the commit message be amended to:
> 
> 1. include upstream sha (`git cherry-pick -x` will do this automatically)
> 2. add Bug: tag
> 3. add signed off by tag (`git commit -s -amend`)

Done for the first two.  The signed-off tag seems to be a kernel process [1].  So I skipped it.  The owner of this AOSP change and also the 'Author' in the git history will report my name in the cherry-pick.  So I think it's ok.

[1] https://stackoverflow.com/questions/1962094/what-is-the-sign-off-feature-in-git-for

Patch-set: 2
Reviewer: Gerrit User 1058860 <1058860@85c56323-6fa9-3386-8a01-6480fb634889>