[UPSTREAM] Fix Calling Convention through aliases
r369697 changed the behavior of stripPointerCasts to no longer include
aliases.  However, the code in CGDeclCXX.cpp's createAtExitStub counted
on the looking through aliases to properly set the calling convention of
a call.

The result of the change was that the calling convention mismatch of the
call would be replaced with a llvm.trap, causing a runtime crash.

Differential Revision: https://reviews.llvm.org/D68584

llvm-svn: 373929
(cherry picked from commit 8a410bcef02c25b6b3a49fa6483875eb64539aad)
Bug: http://b/139945549
Bug: http://b/145013540
Change-Id: I2dafcb7b7353a66e518d77b68b1e98012fa96319