Fix regex_traits<T>::char_class_type for Android.
Android can't use ctype_base::mask for its char_class_type because the
bionic ctype implementation comes from openbsd, which uses an 8-bit
ctype mask which isn't large enough to encode the standard ctypes and
__regex_word.

TODO: Can we drop this? I'm not sure if this is still needed, but
we've been using it since 2014. Also need to determine if dropping
this would be a meaningful ABI break.

Change-Id: Ic6591c23fe788f667eb48d61d2daa95c5d123850