Update patch set 1
Patch Set 1:

> ld.lld: error: linking module flags 'Dwarf Version': IDs have conflicting behaviors in 'out/soong/.intermediates/external/sonivox/arm-wt-22k/libsonivox/android_arm64_armv8-a_cortex-a73_static_cfi_com.android.media/libsonivox.a(eas_mdls.o at 256860)' and 'ld-temp.o'

This is because the new Clang is not able to compile eas_mdls.cc and falling back to the old clang.  So lld is complaining when linking bitcode from two different compilers.

In `clang-error.log` in the failing build, eas_mdls.c is failing with the new clang with:
> external/sonivox/arm-wt-22k/lib_src/eas_mdls.c:1333:9: error: misleading indentation; statement is not part of the previous 'if' [-Werror,-Wmisleading-indentation]

We have a workaround for this scenario (where a new warning flag is added to Clang that the old prebuilt doesn't know).  You can add `-Wno-error=misleading-indentation` to DISABLED_WARNINGS in do_test_compiler.py.

Patch-set: 1
Reviewer: Gerrit User 1058860 <1058860@85c56323-6fa9-3386-8a01-6480fb634889>