[Android] use -l:libunwind.a with -rtlib=compiler-rt
On Android, the unwinder will usually/always be linked statically, but
if -shared-libgcc is passed, then link against -l:libunwind.so instead.

Adjust the -ldl logic slightly. For Android, there is only libgcc.a (and
no libgcc_s or libgcc_eh), so libgcc is always linked statically.
However, passing -static-libgcc would prevent the inclusion of -ldl.
This seems wrong. What we apparently want instead is to add -ldl, except
for static binaries, where the unwinder dl* APIs come from libc.a
instead of libdl.a. (On glibc, the unwinder dl* APIs are in both libc.a
and libc.so. The libdl.so dependency is an Android quirk.)

Test: llvm_android/build.py
Change-Id: I1c98e1b1d6b4f65d0344134a1c4ca8126b6e31b2