[Android][compiler-rt] Adjust build system for -rtlib=compiler-rt
Detecting the default target arch assumes we can link something, but we
can't link anything when we're configuring the builtins, because the
test executable would require the builtins. Instead, use the
COMPILER_RT_DEFAULT_TARGET_TRIPLE variable.

Keep -Wl,-z,defs on the linker command-line, as removing it would hide
unresolved symbols from the builtins and/or the unwinder. Link
libunwind.a into the sanitizer libraries.

Test: llvm_android/build.py
Change-Id: I8fb482a7666d33b73faae60b90e82764f75d432c