SEND UPSTREAM: Add test checks for available fs features.
This isn't quite correct for Android, since we really need to check
this at runtime (it matters which version of the OS the device running
the test has).

Most if not all of these are actually okay for Android when running in
an app context but not when run from the shell in /data/local/tmp,
which is how the tests are currently run on Android.

Change-Id: I125bbe7cba384a0aca9bc1d07eee53797a88c93b